backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3

# Email processing
aiosmtplib>=1.1.6
python-imap>=1.0.0
email-validator==2.1.0.post1

# Audio processing
openai==1.12.0
torch>=2.0.0
torchaudio>=2.0.0
numpy==1.26.4

# Content management
jinja2==3.1.3
pyyaml>=6.0.0
markdown>=3.3.4
python-frontmatter>=1.0.0

# Analysis and search
chromadb==0.4.22
tiktoken>=0.5.0
sentence-transformers==2.3.1
scikit-learn>=1.0.2

# Organization and hierarchy
networkx>=2.6.3
python-slugify==8.0.4

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.98.8
coverage==7.4.3

# Development
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints==2.0.1

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==24.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
python-multipart==0.0.9

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tqdm==4.66.2
requests==2.31.0
urllib3==2.2.0
certifi==2024.2.2
chardet==5.2.0
idna==3.6 